django/requirements.txt
Django==4.2.0
django-cors-headers>=4.2.0
djangorestframework>=3.11,<4.0
psycopg2-binary>=2.8
redis>=4.0
requests>=2.31.0
celery[redis]>=5.4.0
